requirements.txt
requests
pyGitHub
pandas
transformers
setfit
ollama
gradio
python-dotenv
pytest
langchain-ollama